Chore: Rearranged code into relevant files, added missing comments in utility methods
